gst-sdx: Add xrt-dev as run-time dependency
If the xrt packageconfig is used, xrt-dev is needed at run-time, otherwise the
following XRT error occurs while using any of the plugins:

[XRT] ERROR: No devices found
../ext/xrtutils/xcl2.cpp:36 Error calling err = cl::Platform::get(&platforms), error code is: -1

Add dev package to INSANE_SKIP as by default, dev packages should not be
included as run-time dependencies.

While at it, remove sds-lib as run-time dependency for the sds-lib packageconfig
as it is already defined as build dependency and thus included as run-time
dependency by default.