notebooks: Add function to query media device
Add a function that checks the existence of a capture device by querying the
media devices using media-ctl. It returns the media device node if the device
is present.

This also allows to remove the hardware validation check which
checks the xocl text file which is less robust.

In addition, the io-mode property needs to be set on the ghostpad now.