Jenkinsfile: Copy xclbin/xsa from artifacts
If xclbin/xsa are used from build artifacts, copy the files into the local
workspace instead of using symlinks. With symlinks, the time stamp of the files
could be older than the patform in which case make would attempt to rebuild the
xsa/xclbin even though the files exist. Copying the files will update the time
stamp and the files size is small so the impact is negligible.