notebooks: Use dynamic-config property to program coefficients
Update nb6 to use the dynamic-config property to program the kernel coefficients
instead of writing the values into the json file. The json file is meant to be
static and used only during init while the dynamic-config property can be used
to pass a json string to the kernel and set kernel parameters in the kernel
start function.